extract delete_dialog_widget
